Change path that the PyMake repo is mounted at.
This is necessary with the current implementation of the `CallerInfo`
class, which looks for the first stack frame that does not contain
`/pymake/` in the file path. By mounting the PyMake repo at `/pymake`,
all stack frames from any script in the repo was being detected as being
part of PyMake, causing `CallerInfo` to walk all the way up to a built
in Python method's stack frame.